--- event/src/main/java/org/example/event/sourcing/order/poc/event/model/OrderEventName.java
@@ -1,5 +1,0 @@
-package org.example.event.sourcing.order.poc.event.model;
-
-public enum OrderEventName {
-    CREATED, COMPLETED
-}
