--- event/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-org.example.event.sourcing.order.poc.event.config.TopicConfig
