+++ modules/event/src/main/java/org/example/event/sourcing/order/poc/modules/event/model/OrderEventName.java
@@ -1,0 +1,5 @@
+package org.example.event.sourcing.order.poc.modules.event.model;
+
+public enum OrderEventName {
+    CREATED, COMPLETED
+}
