+++ modules/event/src/main/java/org/example/event/sourcing/order/poc/modules/event/model/PaymentEventName.java
@@ -1,4 +1,4 @@
-package org.example.event.sourcing.order.poc.event.model;
+package org.example.event.sourcing.order.poc.modules.event.model;
 
 public enum PaymentEventName {
     CREATED,
