+++ modules/event/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+config.org.example.event.sourcing.order.poc.modules.event.TopicConfig
