+++ order/command-side/src/test/java/org/example/event/sourcing/order/poc/command/order/producer/OrderEventProducerTest.java
@@ -9,7 +9,7 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.TopicPartition;
 import org.assertj.core.api.BDDAssertions;
-import org.example.event.sourcing.order.poc.event.model.OrderEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.OrderEvent;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,8 +21,8 @@ import org.springframework.kafka.support.SendResult;
 import java.time.Instant;
 import java.util.concurrent.CompletableFuture;
 
-import static org.example.event.sourcing.order.poc.event.model.OrderEvent.ORDER_TOPIC;
-import static org.example.event.sourcing.order.poc.event.model.OrderEventName.CREATED;
+import static org.example.event.sourcing.order.poc.modules.event.model.OrderEvent.ORDER_TOPIC;
+import static org.example.event.sourcing.order.poc.modules.event.model.OrderEventName.CREATED;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.mock;
