+++ order/event-handler/build.gradle
@@ -3,7 +3,7 @@ jib.to.image = "noahhsu/espoc-order-${project.name}"
 dependencies {
     implementation project(":common")
     implementation project(":modules:client")
-    implementation project(":event")
+    implementation project(":modules:event")
     implementation project(":modules:observation")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
