+++ order/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/order/domain/OrderEventHandler.java
@@ -1,6 +1,6 @@
 package org.example.event.sourcing.order.poc.handler.order.domain;
 
-import org.example.event.sourcing.order.poc.event.model.OrderEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.OrderEvent;
 
 import java.util.concurrent.CompletableFuture;
