+++ order/query-side/src/main/java/org/example/event/sourcing/order/poc/query/order/domain/entity/OrderEventRecord.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.example.event.sourcing.order.poc.event.model.OrderEventName;
+import org.example.event.sourcing.order.poc.modules.event.model.OrderEventName;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
