+++ order/query-side/src/main/java/org/example/event/sourcing/order/poc/query/order/domain/handler/OrderEventRecordHandler.java
@@ -1,8 +1,6 @@
 package org.example.event.sourcing.order.poc.query.order.domain.handler;
 
-import org.example.event.sourcing.order.poc.event.model.OrderEvent;
-
-import java.util.concurrent.CompletableFuture;
+import org.example.event.sourcing.order.poc.modules.event.model.OrderEvent;
 
 public interface OrderEventRecordHandler {
     void onEvent(OrderEvent event);
