+++ order/query-side/src/main/java/org/example/event/sourcing/order/poc/query/order/domain/handler/OrderRecordHandler.java
@@ -1,6 +1,6 @@
 package org.example.event.sourcing.order.poc.query.order.domain.handler;
 
-import org.example.event.sourcing.order.poc.event.model.OrderEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.OrderEvent;
 
 import java.net.SocketException;
