+++ payment/command-side/build.gradle
@@ -2,7 +2,7 @@ jib.to.image = "noahhsu/espoc-payment-${project.name}"
 
 dependencies {
     implementation project(":common")
-    implementation project(":event")
+    implementation project(":modules:event")
     implementation project(":modules:client")
     implementation project(":modules:observation")
