+++ payment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/payment/consumer/PaymentConsumer.java
@@ -2,14 +2,14 @@ package org.example.event.sourcing.order.poc.handler.payment.consumer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.event.model.PaymentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent;
 import org.example.event.sourcing.order.poc.handler.payment.domain.PaymentEventHandler;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
-import static org.example.event.sourcing.order.poc.event.model.PaymentEvent.PAYMENT_EVENT_HANDLER_GROUP_ID;
-import static org.example.event.sourcing.order.poc.event.model.PaymentEvent.PAYMENT_TOPIC;
+import static org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent.PAYMENT_EVENT_HANDLER_GROUP_ID;
+import static org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent.PAYMENT_TOPIC;
 
 @Component
 @Slf4j
