+++ payment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/payment/domain/PaymentEventHandler.java
@@ -1,6 +1,6 @@
 package org.example.event.sourcing.order.poc.handler.payment.domain;
 
-import org.example.event.sourcing.order.poc.event.model.PaymentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent;
 
 import java.util.concurrent.CompletableFuture;
