+++ payment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/payment/domain/impl/PaymentEventHandlerImpl.java
@@ -2,7 +2,7 @@ package org.example.event.sourcing.order.poc.handler.payment.domain.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.event.model.PaymentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent;
 import org.example.event.sourcing.order.poc.handler.payment.domain.PaymentEventHandler;
 import org.springframework.stereotype.Component;
