+++ payment/query-side/build.gradle
@@ -2,7 +2,7 @@ jib.to.image = "noahhsu/espoc-payment-${project.name}"
 
 dependencies {
     implementation project(":common")
-    implementation project(":event")
+    implementation project(":modules:event")
     implementation project(":modules:observation")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
