+++ payment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/payment/consumer/PaymentConsumer.java
@@ -2,7 +2,7 @@ package org.example.event.sourcing.order.poc.query.payment.consumer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.event.model.PaymentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent;
 import org.example.event.sourcing.order.poc.query.payment.domain.handler.PaymentRecordHandler;
 import org.springframework.kafka.annotation.DltHandler;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -16,8 +16,8 @@ import org.springframework.retry.annotation.Backoff;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import static org.example.event.sourcing.order.poc.event.model.PaymentEvent.PAYMENT_STATUS_GROUP_ID_PREFIX;
-import static org.example.event.sourcing.order.poc.event.model.PaymentEvent.PAYMENT_TOPIC;
+import static org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent.PAYMENT_STATUS_GROUP_ID_PREFIX;
+import static org.example.event.sourcing.order.poc.modules.event.model.PaymentEvent.PAYMENT_TOPIC;
 
 @Component
 @Slf4j
