+++ payment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/payment/domain/entity/PaymentRecord.java
@@ -8,7 +8,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.example.event.sourcing.order.poc.event.model.PaymentEventName;
+import org.example.event.sourcing.order.poc.modules.event.model.PaymentEventName;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.Instant;
