+++ shipment/command-side/src/main/java/org/example/event/sourcing/order/poc/command/shipment/service/ShipmentService.java
@@ -3,14 +3,14 @@ package org.example.event.sourcing.order.poc.command.shipment.service;
 import lombok.RequiredArgsConstructor;
 import org.example.event.sourcing.order.poc.command.shipment.producer.ShipmentEventProducer;
 import org.example.event.sourcing.order.poc.common.model.Shipment;
-import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
-import org.example.event.sourcing.order.poc.event.model.ShipmentEventName;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEventName;
 import org.example.event.sourcing.order.poc.modules.observation.annotation.LogInfo;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 
-import static org.example.event.sourcing.order.poc.event.model.ShipmentEventName.*;
+import static org.example.event.sourcing.order.poc.modules.event.model.ShipmentEventName.*;
 
 @Service
 @RequiredArgsConstructor
