+++ shipment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/shipment/consumer/ShipmentConsumer.java
@@ -2,13 +2,14 @@ package org.example.event.sourcing.order.poc.handler.shipment.consumer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent;
 import org.example.event.sourcing.order.poc.handler.shipment.domain.ShipmentEventHandler;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
-import static org.example.event.sourcing.order.poc.event.model.ShipmentEvent.*;
+import static org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent.SHIPMENT_EVENT_HANDLER_GROUP_ID;
+import static org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent.SHIPMENT_TOPIC;
 
 @Component
 @Slf4j
