+++ shipment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/shipment/domain/ShipmentEventHandler.java
@@ -1,6 +1,6 @@
 package org.example.event.sourcing.order.poc.handler.shipment.domain;
 
-import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent;
 
 import java.util.concurrent.CompletableFuture;
