+++ shipment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/shipment/domain/handler/ShipmentRecordHandler.java
@@ -1,6 +1,6 @@
 package org.example.event.sourcing.order.poc.query.shipment.domain.handler;
 
-import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent;
 
 public interface ShipmentRecordHandler {
