+++ shipment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/shipment/domain/handler/impl/ShipmentRecordHandlerImpl.java
@@ -2,7 +2,7 @@ package org.example.event.sourcing.order.poc.query.shipment.domain.handler.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
+import org.example.event.sourcing.order.poc.modules.event.model.ShipmentEvent;
 import org.example.event.sourcing.order.poc.modules.observation.annotation.LogInfo;
 import org.example.event.sourcing.order.poc.query.shipment.domain.entity.ShipmentRecord;
 import org.example.event.sourcing.order.poc.query.shipment.domain.handler.ShipmentRecordHandler;
